+++ todo/src/App.js
@@ -1,44 +1,13 @@
 import React from "react";
-import { connect } from "react-redux";
-import { inputChange, addTodo, clearCompleted } from "./state/actionCreator";
 import TodoList from "./Components/TodoList";
+import Form from "./Components/Form";
 import "./App.css";
 
-function App({ formValues, inputChange, addTodo, clearCompleted }) {
-  const onChange = e => {
-    inputChange(e);
-  };
-  const onAddTodo = e => {
-    addTodo(formValues);
-  };
-  const onClearCompleted = e => {
-    clearCompleted();
-  };
-
+export default function App() {
   return (
     <div className="App">
-      <form>
-        <input onChange={e => onChange(e)} value={formValues}></input>
-        <button type="button" onClick={onAddTodo}>
-          ADD
-        </button>
-        <button type="button" onClick={onClearCompleted}>
-          CLEAR
-        </button>
-      </form>
+      <Form />
       <TodoList />
     </div>
   );
 }
-
-function mapStateToProps(state) {
-  return {
-    todos: state.todos,
-    formValues: state.formValues
-  };
-}
-export default connect(mapStateToProps, {
-  inputChange,
-  addTodo,
-  clearCompleted
-})(App);
